% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{abo.uk}
\alias{abo.uk}
\title{test for ABO compatibility on UK transplant}
\usage{
abo.uk(dABO = "A", cABO = "A", tier = "B")
}
\arguments{
\item{dABO}{A character from 'A', 'B', 'AB', 'O', for donor ABO group}

\item{cABO}{A character from 'A', 'B', 'AB', 'O', for candidate ABO group}

\item{tier}{A character value for UK transplant candidate's TIER classification (options A and B)}
}
\value{
A logical value T/F
}
\description{
ABO compatibility test between donor and candidate according to TIER classification
}
\examples{
abo.uk(dABO = "A", cABO = "A", tier = "B")
}
