% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{age.diff}
\alias{age.diff}
\title{Donor recipient age difference}
\usage{
age.diff(dage = 60, cage = 50)
}
\arguments{
\item{dage}{A numeric value with donor's age.}

\item{cage}{A numeric value with candidate's age.}
}
\value{
A numeric value.
}
\description{
computes punctuation according to donor-recipient age difference
}
\examples{
age.diff(age = 60, cage = 50)
}
