% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{b.blood}
\alias{b.blood}
\title{blood group B match points}
\usage{
b.blood(dABO = "B", cABO = "O", tier = "B", pts = -1000)
}
\arguments{
\item{dABO}{A character from 'A', 'B', 'AB', 'O'}

\item{cABO}{A character from 'A', 'B', 'AB', 'O'}

\item{tier}{A character value for UK transplant TIER classification}

\item{pts}{A negative value with penalization for B candidates}
}
\value{
A numeric value.
}
\description{
computes penalization when donor's group O and  candidate's group B
}
\examples{
b.blood(dABO = "B", cABO = "O", tier = "B", pts = -1000)
}
