% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{candid_check}
\alias{candid_check}
\title{Validates the Candid file.
Makes sure the header matches the header that a candid file should have.
For each line, call blood group and age checks.
FUNCAO NAO ACABADA}
\usage{
candid_check(csv_file)
}
\arguments{
\item{csv_file}{a .csv file}
}
\value{
A logical value T/F
}
\description{
Validates the Candid file.
Makes sure the header matches the header that a candid file should have.
For each line, call blood group and age checks.
FUNCAO NAO ACABADA
}
