% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cp}
\alias{cp}
\title{Candidates' Color Priority}
\usage{
cp(
  data = candidates,
  q2 = 60,
  q3 = 100,
  cPRA1 = 50,
  cPRA2 = 85,
  check.validity = TRUE
)
}
\arguments{
\item{data}{A data frame with information for candidates' waiting list.}

\item{q2}{A numerical value for the median of candidates' waiting list.}

\item{q3}{A numerical value for the third quartile of candidates' waiting list.}

\item{cPRA1}{A numerical value (env$percentage.maximum - env$percentage.maximum) for the lower cPRA cutoff.}

\item{cPRA2}{A numerical value (env$percentage.maximum - env$percentage.maximum) for the higher cPRA cutoff. cPRA2
must be greater than cPRA1.}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
A data frame with a new column 'cp' (color priority)
}
\description{
Classification of candidates according to waiting list
time on dialysis' quartiles and two cPRA cutoff values.
}
\examples{
cp(data = candidates, q2 = 60, q3 = 100, cPRA1 = 50, cPRA2 = 85)
}
