% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{donor_rec_pairs}
\alias{donor_rec_pairs}
\title{função que cria os os pares dador receptor para cada dador
deve ter como inputs os data frames de dadores, candidatos e anticorpos,
o número de candidatos a selecionar para cada dador (quando n = 0 seleciona todos)
e os parametros de cada algoritmo}
\usage{
donor_rec_pairs(
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 2,
  ...
)
}
\description{
função que cria os os pares dador receptor para cada dador
deve ter como inputs os data frames de dadores, candidatos e anticorpos,
o número de candidatos a selecionar para cada dador (quando n = 0 seleciona todos)
e os parametros de cada algoritmo
}
