% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lima.R
\name{lima_mult_v1}
\alias{lima_mult_v1}
\title{Candidates' selection according to Lima's algorithm for multiple donors}
\usage{
lima_mult_v1(
  iso = TRUE,
  df.donors = donors,
  df.abs = cabs,
  df.candidates = candidates,
  n = 2
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{df.donors}{A data frame containing demographics and medical information
for a poll of donors}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{df.candidates}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates with
color priority classification.}

\item{n}{A positive integer to slice the first candidates.}
}
\value{
An ordered data frame with a column 'cp' (color priority),
'sp', 'hi' and 'mmHLA'.
}
\description{
Ordering of waitlisted candidates for a given donor and
according to to Lima's algorithm.
}
\examples{
lima_mult_v1(iso = TRUE,
df.donors = donors,
df.abs = cabs,
df.candidates = candidates,
n = 2)
}
