% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Candidates' selection according to any algorithm for multiple donors}
\usage{
mult(
  iso = TRUE,
  df.donors = donors,
  df.abs = cabs,
  df.candidates = candidates,
  algorithm = lima,
  n = 2
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{df.donors}{A data frame containing demographics and medical information
for a poll of donors}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{df.candidates}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates with
color priority classification.}

\item{n}{A positive integer to slice the first candidates.}

\item{algoritm}{The name of the function to use}
}
\value{
An ordered data frame with a column 'cp' (color priority),
'sp', 'hi' and 'mmHLA'.
}
\description{
Ordering of waitlisted candidates for a given donor and
according to any algorithm.
}
\examples{
mult(iso = TRUE,
df.donors = donors,
df.abs = cabs,
df.candidates = candidates,
n = 2)
}
