% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lima.R
\name{pre_calculate_lima}
\alias{pre_calculate_lima}
\title{Pre-calculating the order of donors to candidates}
\usage{
pre_calculate_lima(
  iso = TRUE,
  df.donors = donors,
  df.abs = cabs,
  df.candidates = candidates
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{df.donors}{A data frame containing demographics and medical information
for a poll of donors}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{df.candidates}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates with
color priority classification.}
}
\value{
An ordered data frame with a column 'cp' (color priority),
'sp', 'hi' and 'mmHLA'.
}
\description{
Pre-calculating the order of donors to candidates of Lima
}
\examples{
lima_mult_v1(iso = TRUE,
df.donors = donors,
df.abs = cabs,
df.candidates = candidates)
}
