% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts_age}
\alias{pts_age}
\title{Points for age differences}
\usage{
pts_age(donor.age = 60, candidate.age = 40, pts = 4)
}
\arguments{
\item{donor.age}{A numeric value with donor's age.}

\item{candidate.age}{A numeric value with candidate's age.}

\item{pts}{A numerical value for the points to age difference}
}
\value{
A numerical value for pre-defined points
}
\description{
Punctuation given for age difference between candidates and donors
}
\examples{
pts_age(donor.age = 60, candidate.age = 40, pts = 4)
}
