% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts_PRA}
\alias{pts_PRA}
\title{Points for cPRA sensitization}
\usage{
pts_PRA(cPRA = 0, points.80 = 8, points.50 = 4, check.validity = TRUE)
}
\arguments{
\item{cPRA}{Percentual value of cPRA (Between 0 and 100)}

\item{points.80}{A numerical value for the points to a cPRA >= 80}

\item{points.50}{A numerical value for the points to a cPRA >= 50}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
A numerical value for pre-defined points
}
\description{
Punctuation given for sensitized patients according to cPRA value
}
\examples{
pts_PRA(cPRA = 0, points.80 = 8, points.50 = 4)
}
