% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{ric}
\alias{ric}
\title{donor-recipient Risk Index Combination}
\usage{
ric(
  DRI = "D1",
  data = ex.candidates.uk,
  D1R1 = 1000,
  D1R2 = 700,
  D1R3 = 350,
  D1R4 = 0,
  D2R1 = 700,
  D2R2 = 1000,
  D2R3 = 500,
  D2R4 = 350,
  D3R1 = 350,
  D3R2 = 500,
  D3R3 = 1000,
  D3R4 = 700,
  D4R1 = 0,
  D4R2 = 350,
  D4R3 = 700,
  D4R4 = 1000
)
}
\arguments{
\item{DRI}{Donor RisK Index group from options: 'D1','D2','D3','D4'}

\item{data}{A data file with candidates information for UK transplant}

\item{DiRj}{A numeric value (0-1000) for teh combination of indexes}
}
\value{
A tibble with a new column 'ric' that gives the  Risk Index Combination.
}
\description{
computes Risk Index Combination for each pair donor-recipient
}
\examples{
ric(DRI = 'D1', data =ex.candidates.uk,
}
