% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{several}
\alias{several}
\title{Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times to provide statistics}
\usage{
several(iteration.number = 10, should.seed = FALSE, seed = 123, ...)
}
\arguments{
\item{iteration.number}{Number of times the matchability runs.}

\item{should.seed}{Bool to decide whether to set a seed number.}

\item{seed}{Seed for new random number. It's only used when should.seed = TRUE}
}
\value{
Statistics related to all the times the function ran.
}
\description{
Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times to provide statistics
}
\examples{
several(iteration.number = 10,
should.seed = FALSE,
seed = 123,
...)
}
