% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{several}
\alias{several}
\title{Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times to provide statistics}
\usage{
several(iteration.number = 10, should.seed = FALSE, seed = 123, ...)
}
\arguments{
\item{iteration.number}{Number of times the matchability runs.}
}
\value{
Statistics related to all the times the function ran.
}
\description{
Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times to provide statistics
}
\examples{
several(iteration.number = 10,
...)
}
